requirements.txt
asyncio==3.4.3
python-dateutil==2.9.0
python-telegram-bot==21.10
APScheduler==3.11.0
